pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Vaisakhi", email = "dasvaisakhi1@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "flask-cors>=6.0.0",
    "flask>=3.1.1",
    "openai>=1.85.0",
    "python-dotenv>=1.1.0",
    "google-generativeai>=0.8.5",
    "together>=1.5.11",
]
